Restore functional test suite.
This change removes an extra `use` line that was moved out under the
`bldr` component which causes the suite to not compile.

Additionally, the `make functional` Make target was repaired by removing
calls to `cargo test --test function` on components lacking a `tests/`
subdirectory. While this solution is not my first choice, it gets us
back to green without addressing the Makefile for the moment. In other
words, my eyes spot duplication ;)