Fix `bldr start`'s update strategy
Right now, `bldr start` always checks to see if there is software to
download, and always tries to update to the latest version. This commit
adds a flag called `--strategy`, which defaults to `none`. This causes
the supervisor to:

1. Install the software if it is not already on disk
2. Never update the software

This removes many of the current crop of failure conditions. If you pass
the flag `at-once`, then:

1. Install the software if it is not already on disk
2. Try and upgrade the software if it is already on disk and a new
version is available