Rename Bldr crates to Habitat crates.
The following is a large, breaking change to the entire ecosystem and
codebase. Guess what? The project formerly known as Bldr is now Habitat.
Cool, no?

This is the first move at migrating all codebases, toolsets, and related
packages over to Habitat. The Rust crate components change name
here--most interestingly the crate formerly known as 'bldr' is now the
Habitat Supervior, known in code as `habitat_sup` and invoked on the
command line with `hab-sup`. Only the names of the crate projects and
associated `extern crate`, `use`, and error types have been changed
over, but this leaves us with a building system.

The following is a summary of the crate changes as they appear now, in
pseudocode:

```
habitat_core = {
  directory: core
  crate_lib: habitat_core
}

habitat_depot_core = {
  directory: depot-core
  crate_lib: habitat_depot_core
}

habitat_depot_client = {
  directory: depot-client
  crate_lib: habitat_depot_client
}

habitat_depot = {
  directory: depot
  crate_lib: habitat_depot
  crate_bin: hab-depot
}

habitat_sup = {
  directory: sup
  crate_lib: habitat_sup
  crate_bin: hab-sup
}
```

The unit tests pass, the docs generate, but the functional suite is not
entirely passing due to changes needed in Studio. Here is a current
summary:

```
failures:
    bldr::gossip::incarnation_updates_on_suspicion
    bldr::gossip::isolated_members_find_a_way_to_rejoin
    bldr::gossip::members_are_gossiped
    bldr::gossip::ressurection_of_permanent_members
    bldr::gossip::routes_around_failure
    bldr::gossip::two_supervisors_detect_failure
    bldr::gossip::two_supervisors_link
    bldr::topology::leader::elects_a_leader
    bldr::topology::leader::minimum_quorum

test result: FAILED. 13 passed; 9 failed; 8 ignored; 0 measured
```