Add `TARGET` metadata for all built artifacts.
This change introduces a new piece of metadata: the `TARGET` for a
package. It is similar in nature to Rust's target triple which comes
from the clang's target triple, however we only require 2 pieces of
information:

* The target architecture - such as `x86_64`, `i386`, `armv7`, etc.
* The target system - maps to a generic concept of an operating system,
  although more specifically the target kernel such as `linux`,
  `darwin`, `freebsd`, etc.

At build time, the architecture is derived by lower-casing the result of
`uname -m` and the system by lower-caing the result of `uname -s`.
Combined together, this is known as the `TARGET` for this package. That
is, the kernel/architecture combination that will run the package.

The `hab-build-plan` program has been augmented with the following new
build variables which are set and meant to be optionally consumed by
Plan authors:

* `$pkg_arch` - the target architecture
* `$pkg_sys` - the target system
* `$pkg_target` - the architecture and system values, seperated with a
  dash. (ex: "x86_64-linux", "i386_openbsd", etc.)

The artifact files that are produced will contain the target at the end
of the filename
(ex: happyhumans-possums-8.1.4-20160427165340-x86_64-linux.hart) which
should help if a system ends up with many artifact files in the future
destined for multiple targets.

The rest of the codebase has been updated to *only* expect the
`x86_64-linux` target which has been the only supported execution
environment to date. Once a second target is supported, the installation
code as well as the Depot code will need to become target-aware.
However, we should be able to augment and support new targets in the
future and ignore this metadata until then.

References:

* http://clang.llvm.org/docs/CrossCompilation.html#target-triple
* https://github.com/rust-lang/rust/tree/master/src/librustc_back/target
* http://linux.die.net/man/1/uname

Pull request: #449
Approved by: adamhjk