Refactor core crypto module.
This change performs a refactoring of the encryption and cryptographic
functions in the `core` component. In the last 1-2 months much of the
underlying implementation has been swapped out (a GPG-based system to
libsodium) and much more functionality has been added. As we seem to be
approaching a steady state for at least the near-term it was worth
revisiting what we've built up, and we noticed 3 issues:

1. The size of the single module file was racing upwards with new
functionality
2. The number of tests were not keeping pace and worse, to add them led
to an even larger single file that was getting harder to navigate
3. The call sites for the crypto code often lacked the context to
determine the key cache path and some crypto functions required this
even though it didn't use it at all

This new code organization splits the main module into several smaller
ones:

* `crypto::artifact` which contains behavior related to Habitat artifact
  files
* `crypto::hash` which contains behavior relating to hashing (sometimes
  called checksumming) files and artifacts
* `crypto::keys` which contains several sub-modules relating to the 3
  primary key types: Sym keys (used to encrypt communication), Box keys
  (used to encrypt/decrypt files and payloads destined for service
  groups), and Sig keys (used to sign/verify artifacts)

The updated key-related APIs have been reorganized in an attempt to be
consisten with one another and to not leak their internal implementation
details. In particular, a consumer of `core::crypto` should **never**
have to explictly import any libsodium modules, structs, or functions.
Keys are type aliases of a generic type which is maintained internally
and all specific public and secret keys are internal to that "key pair"
struct--even if only one of the two keys are present or relevant. When
working with keys, only the key pair struct are passed around.

In addition, a number of unit tests have been added so that the public
interfaces of each module has good and reasonable coverage that
exercises the happy path and exceptional cases.

The `crypto::hash` test suite has a first attempt at an explicit
functional test which is activated by enabling a feature called
"functional" when running tests: `cargo test --features functional`.
This test in particular downloads a large file from a website in order
to verify its hash/checksum without commiting a multi-megabyte file into
the codebase. Bonus points for trying out hyper's http_proxy support
which (eventually) worked like a charm ;)

Finally, the command-related code in the `hab` component was
re-organized so that every subcommand's "business logic" function lived
in a consistent module space that tightly tracks its own dependencies
and any related internal functions.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #502
Approved by: metadave