[hab,hab-sup,hab-director] Add http & https proxy support.
This change re-introduces support for using HTTP proxies when contacting
a Depot service. Previous versions of the hyper Rust crate were not
sufficiently flexible to support both custom SSL contexts and proxy
support which led to the removal of proxy support in #652.

This new strategy adds support back by using a custom
Hyper `NetworkConnector` that supports plaintext (i.e. http) and
tunneled (i.e. https) proxying modes as well as optional basic
authorization support. All 4 variants should cover the vast majority of
corporate network/proxy setups.

Background Reading Materials
----------------------------

Several references, documents, and RFCs were used in working up this
feature and are shared here for context and for the curious:

* RFC 2616 - "Hypertext Transfer Protocol -- HTTP/1.1"
  https://tools.ietf.org/html/rfc2616
* RFC 2617 - "HTTP Authentication: Basic and Digest Access Authentication"
  https://tools.ietf.org/html/rfc2617
* RFC 2817 - "Upgrading to TLS Within HTTP/1.1"
  https://tools.ietf.org/html/rfc2817
* RFC draft - "Tunneling TCP based protocols through Web proxy servers"
  http://www.web-cache.com/Writings/Internet-Drafts/draft-luotonen-web-proxy-tunneling-01.txt
* Polipo Manual - "Access Control"
  https://www.irif.univ-paris-diderot.fr/~jch/software/polipo/polipo.html#index-authCredentials
* Squid Cache Wiki - "Feature: HTTPS (HTTP Secure or HTTP over SSL/TLS)"
  http://wiki.squid-cache.org/Features/HTTPS
* curl Man Page - "Environment"
  https://curl.haxx.se/docs/manpage.html
* Wget Manual - "Proxies"
  https://www.gnu.org/software/wget/manual/html_node/Proxies.html
* Arch Linux Wiki - "Proxy settings"
  https://wiki.archlinux.org/index.php/proxy_settings

Environment Variable Behavior
-----------------------------

* `http_proxy` / `HTTP_PROXY` - A URL for a proxy server that will be
  used for any HTTP connections. You may specify a username and password
  through the proxy URL to support `Basic` authorization. For example:
  `"http://user:pass@proxy.example.com:8123"`. The lowercase version of
  this environment variable takes precedence over the uppercase version.
* `https_proxy` / `HTTPS_PROXY` - A URL for a proxy server that will be
  used for any HTTPS connections. This uses a TCP tunneling mode,
  meaning that the proxy server tunnels the encrypted socket connection
  through to the target and has no ability to intercept or unpack the
  contents of the connection. You may specify a username and password
  through the proxy URL to support `Basic` authorization. For example:
  `"http://user:pass@proxy.example.com:8123"`. The lowercase version of
  this environment varible takes precedence over the uppercase version.
* `no_proxy` / `NO_PROXY` - A comma-separated list of domain extensions
  that a proxy should *not* be used for. For example, if the value of
  `no_proxy` is `".example.com"` the a proxy should not be used for URLs
  ending with the `example.com` domain. The lowercase version of this
  environment variable takes precedence over the uppercase version.

Bonus: ApiClient
----------------

In additional to supporting the above feature a refactoring of the
habitat_http_client component led to the creation of an "API Client"
struct to wrap up the underyling hyper HTTP behavior. It centers around
a common base URL so that a proxy server can be determined and set up
correctly for the target. As proxy authentication is implemented quite
differently between plaintext and tunneled mode, it was easier to
encapsulate this behavior inside the `ApiClient`. See the developer
documentation for more usage details.

Bonus: User-Agent Headers
-------------------------

Finally, the wrapping of HTTP client calls behind the above mentioned
`ApiClient` allows us to include a common `User-Agent` HTTP header which
can be used to measure the various versions of software that connect to
a Depot or other endpoints. The format of the `User-Agent` header value
is similar to the analytics code implementation in the `hab` CLI.

Final Word About depot-client API
---------------------------------

Supporting an injection of a product name and version into each Depot
client has led to the addition of function parameters in the codebase
where a client is created. This isn't yet totally ideal, but future work
is slated to reduce the number of Depot client creations and to simplfy
its setup API. The author of this current feature asks for forgiveness
in the meantime and not permission ;)

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>