Use Cargo "workspaces" to manage a project-wide Cargo.lock file
There is now a Cargo.toml and Cargo.lock file at the root of the Git project. This is what the
Cargo RFC refers to as a "Virutal Manifest". This references all other crates and they reference
back to the virtual manifest. All crate targets are now built to $root/targets instead of a
target directory in their respective crate.

Workspaces were officially added to Cargo in version 0.12 which will be included in Rust 1.11. You
can, and should, download Cargo 0.12 or greater to compile our project. For more information about
how to use workspaces see the docs or RFC:

* [Documentation](https://github.com/rust-lang/rfcs/blob/master/text/1525-cargo-workspace.md)
* [RFC](https://github.com/rust-lang/rfcs/blob/master/text/1525-cargo-workspace.md)

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>