[hab,sup] Add `UI#progress()` method to control progress bar behavior.
This change adds a `progress()` method on the `UI` struct to determine
whether or not to display a progress bar on download/upload activities.
The decision is based on whether or not the stdout stream is a terminal
(i.e. an interactive session) meaning that progress bars will be removed
in non-interactive contexts.

Here is an example of the previous behavior (minus coloring which cannot
be reproduced here):

```
> /src/target/debug/hab install core/jq-static
» Installing core/jq-static
↓ Downloading core/jq-static/1.5/20160730001211
    828.85 KB / 828.85 KB \ [====================================================] 100.00 % 462.36 KB/s
↓ Downloading core-20160612031944 public origin key
    75 B / 75 B | [==============================================================] 100.00 % 824.83 KB/s
☑ Cached core-20160612031944 public origin key
✓ Installed core/jq-static/1.5/20160730001211
★ Install of core/jq-static/1.5/20160730001211 complete with 1 new packages installed.
```

Finally, here is the same command whose output stream is attached to a
unix pipe:

```
> /src/target/debug/hab install core/jq-static | cat
» Installing core/jq-static
↓ Downloading core/jq-static/1.5/20160730001211
↓ Downloading core-20160612031944 public origin key
☑ Cached core-20160612031944 public origin key
✓ Installed core/jq-static/1.5/20160730001211
★ Install of core/jq-static/1.5/20160730001211 complete with 1 new packages installed.
```

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>