Static directory permissions
This commit adds the static directory to the default directory
structure of services in /hab/svc and it ensures the ownership is set
consistently to avoid permissions errors.

We also need the static directory copied in the api application
instead of linked because the parent (/hab/svc/hab-builder-api) will
be owned by root, which creates a permissions issue for the init hook
to continue running and set up the habitat.conf.js config file.

Signed-off-by: jtimberman <joshua@chef.io>