Improve client & server error reporting
This change will attempt to convert non-200 HTTP responses from the
Depot and API servers into a NetError which now implements
std::error::Error and is bubbled up back to the `hab` crate. These
errors contain a code and a message to help identify what exactly
went wrong with the users remote request.

* protocol::net::NetError now implements std::error::Error
* Added helper function for converting ErrCode into an HTTP status
* Rename depot_client::Error::HTTP to depot_client::Error::APIError and
  include status and error message in error
* Attempt to parse and display net errors returned from API server
  in builder-depot-client
* Added additional error handling to provide more granular errors
  from api server
* Remove ErrCode::INTERNAL - ErrCode::BUG is the generic "that was bad"
  error code

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>