Drop Cargo nightly requirement as of Rust 1.12.0 release.
This change was possible due to the release of Rust 1.12.0 which
features a version of Cargo which supports the workspace feature in use
in this project.

In an effort to minimize and simplify internal setup and external
requirements, we can switch back to the Rust stable version of Cargo.

**Note** that currently the TravsCI infrastructure was left unchanged
until there is good evidence that Rust 1.12.0 is being installed across
their fleet.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>