Improve Builder versioning & release process
There are two major changes in this commit to how we version and
release Builder components. These changes are in the spirit of
allowing us to release the Builder services more rapidly than the
Habitat client components. Since our client and server do not need to
release in lock-step, we can (and should), release Builder as often
as possible.

* Use monotonically increasing version number when building with
  hab-plan-build (release builds)
* Use git commit hash for version when building with Cargo (dev builds)

In addition to these major changes, this commit also removes the guide
for releasing Builder and replaces it with a shell script to allow an
engineer to manually build and publish all of the builder components on
demand.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>