Add builder-db component
This commit adds a builder-db component. It serves as the base layer for
using PostgreSQL for persistence for builder services. Currently, it
provides 3 things:

* A simple `Pool`, which governs access to the database
* A `Migrator`, which handles database migrations
* A `test` namespace, which exports macros and functionality neccessary
  to build integration tests for services which use access the database

The top level module documentation includes some advice on design and
usage of the database within builder.

To run the test suite, you must first, in a separate terminal:

```sh
$ mkdir -p /hab/svc/postgresql
$ cp components/builder-db/tests/db/* /hab/svc/postgresql
$ hab start core/postgresql
```

You can then use `cargo test` as normal.

As we gain more experience writing the service components, I expect we
will automate that part of the process - hence, I'm leaving it out of
the formal documentation for now.

Signed-off-by: Adam Jacob <adam@chef.io>