Rename the vault to the originsrv
It became clear in the process of refactoring that the vault was
ill-named - in truth, it's the `originsrv` - the service responsible for
supplying data that's related to a given origin. This PR renames the
service.

Additionally, it makes the `make bldr-run` command work again. It does
this by firing up a postgresql instance within hab-sup, and creating the
default databases needed by the services.

A word of caution - this will re-configure your database instance.

Another word of caution - this will conflict with running the
integration tests for components with a database.

Signed-off-by: Adam Jacob <adam@chef.io>