Refactor Supervisor main loop to tick over each Service on each tick
This refactor moves much of the logic of the manager's main loop into
a `tick()` function on the Service struct. This function is designed
to return quickly and encapsulate tending to a service on each tick
of the manager. Creating this abstraction allows us to better
represent the execution flow of the supervisor. The Service struct
now owns everything it needs to tend to itself on instruction by
the manager.

Additional changes and improvements:

* Remove `config_from` from global supervisor config. This is now
  part of the `ServiceSpec`/`Service` module
* Better validation of parsing Bindings
* Removed `habitat_sup::package::Package` in favor of
  `habitat_core::package::PackageInstall`
* Improve ergonomics of a number of APIs dealing with file paths
  * Use `AsRef<Path>` over `&Path` where possible
  * Use `PathBuf` over `String` where possible
* Move Supervisor specifics from core crate into supervisor crate
  * Added `habitat_sup::fs` module
  * Move `pkg_svc_*` functions from `PackageInstall` struct to
    `habitat_sup::fs` module

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>