Update `toml` to master (as of d39c3f7b3ec95cb3cc1e579d7d747206c66aab74).
We hit the bug described in https://github.com/alexcrichton/toml-rs/issues/145 when our
package struct in the ServiceConfig attempted to serialize exports (which is a HashMap) and
exposes (which is a Vec<String>).  This was resolved by https://github.com/alexcrichton/toml-rs/commit/fba7778782193f0d0899f82d6b322da59a9044a0.

Once the `toml` crate releases an update, we can revert this commit.

Signed-off-by: Steven Murawski <steven.murawski@gmail.com>