[builder] Update version computing logic in plan.sh files.
This change updates how the Builder components calculate their
`pkg_version` value. The resulting behavior is the same, but additional
work needs to happen with the newly introduced `$SRC_PATH` and
`$CACHE_ROOT`. As this method of lazily computing the `pkg_version`
isn't yet officially supported, I'm not concerned with this fix.
However, the logic to deal with changing a `pkg_version` value
mid-flight has been extracted into a common, generic-ish function in
each Plan. This is fast becoming the proto-implementation of a generic
user-facing helper which a Plan author can use for exactly this use
case. A future pull request will add this capability and at that point,
we can delete the manual versions. Moving forward!

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>