Add an endpoint for getting jobs from a project
Here, we add a REST API endpoint for retrieving the 50 most
recently-created jobs for a given project.

In doing this, we update the notion of a "job" in our database, by
adding the following fields:

* `created_at`: the time the job was entered into the system
* `build_start`: the time a `hab studio build` command starts
* `build_stop`: the time the build finishes, successful or not
* `package_ident`: the full identifier of the package generated by the
  job. Currently, this is only set for successful jobs.

These new data are all gathered by the build runner, and are stored in
the jobs database at the end of a build run.

The `set_job_state_v1` database function replaced with an
`update_job_v1` function to account for these additional fields, as was
the JSON serialization logic for `Job` messages.

Additionally, `ProjectJobsGet` and `ProjectJobsGetResponse` protocol
messages were created.

To deal with timestamps being stored and retrieved from the database,
we add a dependency on the `chrono` time library.

Documentation of the endpoint was added to the Builder API RAML file.

Fixes #2256

Signed-off-by: Christopher Maier <cmaier@chef.io>