Remove piped log output for studio teardown
This may or may not be a permanent change. At the moment, we are
investigating possible reasons for the worker hanging during the
teardown phase in our acceptance environment.

Long term, we may add this back in, as one of multiple log streams for
a job (e.g., main job, teardown, or STDOUT / STDERR for each, etc.)

Signed-off-by: Christopher Maier <cmaier@chef.io>