Use ANSI color codes in build output
This will allow us to display the colors in the UI. We are not adding
it to builds running in debug mode, since that presumably will be only
under specific circumstances when we're debugging a problem and ANSI
codes would be more of a nuissance.

A `strip` query parameter is added to the `jobs/{id}/log` API
endpoint. Specifying this parameter with a value of `true` (or
anything that can reasonably be interpreted as `true`; see the RAML
docs for details) will return log content with ANSI codes stripped
out. Leaving the parameter off (or with a `false` value) will yield
log lines with embedded ANSI codes.

Signed-off-by: Christopher Maier <cmaier@chef.io>