Add a Build Output tab to the manifest view
This exposes historical logs by adding a Build Output tab to the manifest view, visible to signed-in users. It does this by pulling the build-log view into a new component shared by both the in-progress ("streaming") build view and the manifest view.  Also adds a bit of cleanup by formatting release dates in lists, and fixes a date-display bug in recent-packages list on the dashboard.

Signed-off-by: Christian Nunciato <cnunciato@chef.io>