Tweak archive configuration
Now, you specify which backend you want by providing a `backend` parameter, like so:

```toml
[archive]
backend = "s3"
key = "..."
secret = "..."
...
```

or

```toml
[archive]
backend = "local"
local_dir = "/path/to/my/archive"
```

The `backend` key is required; all others depend on the specific
backend you're configuring. This structure ought to make it a bit
easier to work with.

To tighten things up, the backends are defined by an Enum; in order to
do this, I had to update `toml` from `0.4.0` to `0.4.1` to get the
correct deserialization logic.

This also adds configuration templating for archive information. Turns
out you need this; who knew? ;)

Finally, it adds a default to store logs locally in `/tmp`. You don't
really want to do this in real life, of course.

Signed-off-by: Christopher Maier <cmaier@chef.io>