Use correct default values for log archiver
This effectively undoes commit d0345238, but fixes the config to how
it should have been in the first place. TL;DR: I wrote up the correct
default values in our `BUILDER_DEV.md` file, but made a typo in the
actual `default.toml` file. The configuration value is `local_dir`,
not `dir`.

The default values in the above-referenced commit cause problems when
configuring the archiver for S3, in that the `endpoint` really should
default to `None`. This value is only needed if you need to use path
style URLs for your S3 objects instead of virtual-host style
URLs (e.g. `mybucket.s3.amazonaws.com/my-object`). This is necessary
if you use Minio, though, because it doesn't yet support virtual-host
style URLs. With an empty string as an endpoint default, you always
need to supply an endpoint; otherwise, the archiver does not start up
because an empty string does not parse as a valid URL.

Signed-off-by: Christopher Maier <cmaier@chef.io>