Write out a pre_build.env file before a build happens.
The purpose of this file is to provide metadata to the builder worker if
a build should fail. In such a case, a worker can read and parse this file to
assign version and release information to the ident of the job that failed.

Also, it turns out that project updates didn't actually work in builder-api, so
this fixes that as well.

Signed-off-by: Josh Black <raskchanky@gmail.com>