Refactoring builder services
This change starts the work to eliminate the dependency on a user auth token
for the builder worker service. It modifies some of the internal APIs to not
require a session id. It also adds channel support to the Job object, and
moves part of the promotion code from the scheduler to the worker.

A couple of features are eliminated (for now - some code preserved):
1) the ability to group promote automatically by the scheduler
2) the ability to upload and promote based on a TOML file

Signed-off-by: Salim Alam <salam@chef.io>