Process GitHub Push Hook notifications
This commit changes how we make authenticated requests to the GitHub API
on behalf of the user and adds the ability to process GitHub Webhook
notifications for Ping and Push.

In order to support Ping and Push notifications for Private repositories
we needed to switch to integrating with GitHub using a full GitHub App
instead of just an OAuth App. This PR changes the github-api-client to
only make requests which a GitHub App could make and signs the commits
appropriately.

The authorization process has changed to a per repo basis and no longer
requires a user to setup a personal access token.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>