Schedule a registration retry OnError from RouteSrv
The current behaviour is to shutdown a backend service connected
to RouteSrv if RouteSrv sends a registration conflict error. This
happens when you attempt to register to the RouteSrv and it rejects
your registration.

This change keeps the server running and replies to a REG_NOT_FOUND
with a registration message

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>