Fix RouteClient socket state issue when routing to self
Changes the REQ socket in RouteClient to a DEALER socket to allow
for unrestricted send/recv pattern. This is required when a server
is handling a message and then attempts to route a message within
it's handler to a server and the destination happens to be itself.

Also removed `route_async` from RouteClient. Now that the socket
state is no longer enforced by ZMQ itself, we can't ensure that
messages we receive are completions for a given transaction. We
will need to create a TXN-id, store pending transactions, and then
use the TXN-id to complete those stored transactions. This work
will need to be done a bit later.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>