Add key generation to Builder UI
This adds the ability to generate a new origin key pair when creating an origin
and at any time later, and allows the signing key to be downloaded by members of
the origin.

It also refactors to remove the keys-list component, which was being used for both
public and private keys. (Given how different they are, it doesn't make sense to
cram both into the same component.)

Signed-off-by: Christian Nunciato <cnunciato@chef.io>