[builder-worker] Refactor runner in preparation for exporters.
This change is primarily structural refactorings with the essential
behavior preserved, aside from an explicit crash when the config
`auto_token` is empty or not present.

Among some of the changes:

* Compile in the exact Studio package dependency when building under
Habitat, otherwise fall back to look for the latest installed
`core/hab-studio`. The command resolving is in a `lazy_static!` and so
is invariant once the program starts, no matter what other Studio
packages may be installed in the future.
* Reference Habitat-known environment variables from the relevant
constants.
* Normalize some of the logging output in the `runner` module.
* Extract Studio-spawning code into a `runner::studio` module. Future
exporter work will follow this pattern.
* Set `log_path` to `{{pkg.svc_var_path}}` when running in a package

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>