Finish plan linkage and Docker settings submission
This change adds support for linking new plans both by origin and package, and allows submission of additional Docker settings like org, repo and tags.

It also removes a bunch of dead code and moves the project-specific plan-selection code into a new component inspired by hab-plan-select.

Still some refactoring and consolidation to be done here (as always), but it’s ready for testing and feedback collecting.

Signed-off-by: Christian Nunciato <cnunciato@chef.io>