Store Session in memory instead of database
This change will keep all sessions in memory stored on the heap
instead of in a database row. Sessions should be considered transient
data which exists until the point of expiration or until the server
is restarted, whichever comes first.

We were previously storing sessions in memory in Redis so there was
no reason to store them in memory in the application server. This is
no longer the case with PostgreSQL being the backend for sessionsrv.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>