Use REQ socket for RouteClient - fix reply ordering
Allocate and use a REQ socket when a backend server uses a RouteConn
to create a request to be sent to another backend server. This should
fix the issue where messages would be sent and the message received
would have been intended for a different thread. There may be a path
involving less allocations (where we pre-allocated the RouteClient for
the request) but this is completely sufficient for the time being.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>