Remove builder-web package and move it's contents into builder-api-proxy
Builder-Web can't be run as a service, so it can't be configured or bound
to which means it can only be consumed as a library dependency.

Additionally, we need to pull latest of builder-web regardless of which
channel we're building against. We don't want to consume it as a package
dep because we'll always pull the stable version of it.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>