Integration tests for the Builder HTTP API
Resolves https://github.com/habitat-sh/builder/issues/77

We had a suite of integration tests for the Builder API but it had some
problems. It destroyed the existing PG installation, it was slow to
start up and tear down, and it didn't cover many endpoints.

This aims to rectify these problems as much as possible. It uses the
builder dev studio environment and cleans up after itself when it exits.
It covers as many endpoints as possible. It provides a way to run all
of this stuff in an automated way and works both locally and on travis.

It's also worth noting that we converted the tests from TypeScript to
JavaScript since we weren't using any of the features of TypeScript and
it was just another layer of indirection to figure out when tests
failed.

Signed-off-by: Josh Black <raskchanky@gmail.com>
Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>