[tools] Update the `ssh_helpers/airlock.sh` test.
This update will check to see if the network interface has been placed
in the correct network namespace and has an IP address. Otherwise, it
will fall back to check for the appropriate failure scenarios to report
back to the operator. Note that not all failure scenarios are easily
actionable with an automated script (hence the report messaging).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>