[builder-worker] Re-work job streaming & introduce log sections.
This change consists of 2 changes which are lightly related but more
coupled in implementation (hence one commit).

Re-implement Job Streaming In `JobStreamer`
-------------------------------------------

A new implementation of job streaming was written to replace the
`LogPipe` struct, which is called a `JobStreamer`. Under the covers it
uses the same zeromq-based code but its stream handling and API is
different enough to warrent a new name.

A `JobStreamer` contains a clonable `StreamTarget` which is managed via
a `Mutex` and is the struct that contains the socket and line count
state for the stream connection. The `Arc`-clonable type of the target
allows it to be shared among multiple threads which may be concurrently
attempting to write log lines to the single log stream. As the log
stream is not read from, but rather written to, a `Mutex` seemed to be
the natural locking mechanism. Additional care was taken to prevent a
mutex deadlock by keeping calls to the stream target short and scoped
small.

The primary consuming API is a `consume_child()` function which takes a
mutable reference to a `Child` process. This method will consume both
the process' `stdout` and `stderr` streams and attempts to keep the
output ordering as close to original as possible. Additionally, there is
a `println_stderr()` function which can be used to print error messaging
to the log stream if a particular part of the build goes wrong (i.e.
during a repo clone, downloading origin keys, etc.).

Introduce Log Sections
----------------------

In addition to the `JobStreamer` work, this change introduces a more
formal notion of sections within a log stream. Future work may introduce
new sections, remove others, or possibly change the section delimiting
format, but this is a first step with 2 goals in mind:

1. Reduce the visual importance of the `--- BEGIN: Studio build ---`
lines that were starting to overwhelm the log output
2. Introduce explict start and ends to every section started into,
whether or not the section resulted in a failure. Additionally an
explicit start and end of the log was added to allow a user to know when
the log was indeed finished.

The section names are user-facing one-word tokens which should help
convey what is going on without being too vague. This is why the
previous `"POSTPROCESSING"` part of the log is now a `"publish_package"`
section name.

Closes #93

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>