Create the tmuxinator directory if it doesn't already exist
Also, clean up shellcheck lints,make the script run correctly regardless of the
current directory, and provide some instructive output if run without args.

I removed the "TODO: constrain the list of possible environments?" comment since
I think that is more the domain of the generate-habitat-env-tmuxinator.sh
script. If we added new environments, we don't want to have to modify this.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>