Switch to DogStatsD client
This swaps out the generic StatsD client we were using for a DogStatsD
client. We use DataDog as our metrics aggregator, and this will allow
us to take advantage of some DataDog extensions, particularly metric
tags.

No tags are added in this PR, though; this is just a pure replacement
of the stats client.

Signed-off-by: Christopher Maier <cmaier@chef.io>