Remove generated code from git
I am removing this after realizing that we are all checking in different version of code generated by different versions of protoc. After this change, the build will be the single source of truth for all protocol changes. Also, I have removed the need to run `cargo build --features protocols` as it now runs automatically.

Signed-off-by: Elliott Davis <elliott@excellent.io>