[BREAKING/depot] Updates due to PackageTarget changes in habitat-sh/core#37.
Breaking config.toml update
---------------------------

There is a potential user-impacting breaking change included with this
change. That is, prior to this change an operator of a
`habitat/builder-api` service would override the default `targets`
section of their configuration with a form similar to this:

```toml

[[targets]]
architecture = "x86_64"
platform = "windows"

[[targets]]
architecture = "x86_64"
platform = "linux"
```

(Note that the above example is *exactly* the defaults for the service
and therefore it would be extremely unlikely to see this in the wild)

With this change, an operator would want to update their configuration
with a form similar to this:

```toml

targets = ["x86_64-linux", "x86_64-windows"]
```

Scope of breaking change
------------------------

Given how our software is built and configured, there are likely **no**
impacted users, including the Habitat team operating the public Builder
service. Overriding the default supported package targets is likely
under-documented and would require a very unusual use case to even
warrant this. However, the chance of impacting a running service is
non-zero, hence why this is called out.

---

Here is a high level summary of all the other resulting changes:

* Exploit the newly available iterators on `PackageIdent` and
`PackageTarget` and the `PackageIdent.archive_name_with_target()`
implementation to generate the same S3 key strings in `s3_key()`. The
behavior of `s3_key()` didn't require internal knowledge of a struct,
hence it was moved into a private module function where adding unit
tests was made easier too.
* Update `s3_key()` to return a `Result<String>` rather than a `PathBuf`
so that its successful return could be used directly as a key and so
that any relevant failures could be tracked with Rust error propagation.
* Update the default `Config` to use the newly available supported
`PackageTarget` constants for `x86_64-linux` and `x86_64-windows`.
* Update the `DepotUtil::archive_name()` trait function signature to
operate directly with `PackageIdent`s rather than with generic
`Identifiable`s to gain access to some of `PackageIdent`'s behavior. So
rather than using generics, this change uses type conversion from a
`&OriginPackage` to a `PackageIdent` and from a `&OriginPackageIdent`
to a `PackageIdent`. Several unit tests were added to test the type
conversions.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>