Remove sharding from the database code.
This introduces several new functions to replace functions that were
previously iterating over all of the shards. It removes sharding from
the database queries that run, but leaves it in the config for each
service, so that routing continues to work. It also introduces a new
bash script to collapse all the existing shards into the public schema.

Signed-off-by: Josh Black <raskchanky@gmail.com>