Add memcached support for builder
When a package gets uploaded, promoted or marked as public the latest version of that package gets added to memcached.
When a packages gets demoted or marked as private, it is removed from the cache.

Any calls to `/latest` endpoints will no check the cache to see if the package exists. If it does, it's returned, if not, we round trip to the db for a result and update the cache.

Signed-off-by: Elliott Davis <elliott@excellent.io>